week15/grpc_demo/pyproject.toml
[tool.poetry]
name = "grpc-demo"
version = "0.1.0"
description = ""
authors = ["Patrick Stünkel <webminz@googlemail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio-tools = "^1.71.0"
grpcio = "^1.71.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#[tool.poetry.scripts]
#codegen = "python -m grpc_tools.protoc -I./protos --python_out=grpc_demo --grpc_python_out=grpc_demo --pyi_out=grpc_demo protos/schema.proto"
